Move Sulu Version CompilerPass to AdminBundle to avoid depending on WebsiteBundle parameters
