+++ src/Sulu/Bundle/AdminBundle/DependencyInjection/Compiler/SuluVersionPass.php
@@ -9,72 +9,81 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Component\Util;
+namespace Sulu\Bundle\AdminBundle\DependencyInjection\Compiler;
 
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Finder\SplFileInfo;
 
 /**
+ * @internal
+ *
  * Read versions from composer.lock and composer.json.
  */
 class SuluVersionPass implements CompilerPassInterface
 {
+    private const SULU_DEFAULT_VERSION = '_._._';
+
     public function process(ContainerBuilder $container)
     {
         $dir = \realpath($container->getParameter('kernel.project_dir'));
 
-        $container->setParameter('sulu.version', $this->getSuluVersion($dir));
-        $container->setParameter('app.version', $this->getAppVersion($dir));
+        $container->setParameter('sulu.version', \is_string($dir) ? $this->getSuluVersion($dir) : self::SULU_DEFAULT_VERSION);
+        $container->setParameter('app.version', \is_string($dir) ? $this->getAppVersion($dir) : null);
     }
 
     /**
      * Read composer.lock file and return version of sulu.
-     *
-     * @param string $dir
-     *
-     * @return string
      */
-    private function getSuluVersion($dir)
+    private function getSuluVersion(string $dir): string
     {
-        $version = '_._._';
-
-        /** @var SplFileInfo $composerFile */
         $composerFile = new SplFileInfo($dir . '/composer.lock', '', '');
         if (!$composerFile->isFile()) {
-            return $version;
+            return self::SULU_DEFAULT_VERSION;
         }
 
         $composer = \json_decode($composerFile->getContents(), true);
-        foreach ($composer['packages'] as $package) {
-            if ('sulu/sulu' === $package['name']) {
-                return $package['version'];
-            }
+
+        if (!\is_array($composer)
+            || !\is_array($composer['packages'] ?? null)
+        ) {
+            return self::SULU_DEFAULT_VERSION;
         }
 
-        return $version;
+        foreach ($composer['packages'] as $package) {
+            if (!\is_array($package)
+                || 'sulu/sulu' !== ($package['name'] ?? null)
+            ) {
+                continue;
+            }
+
+            if (!\is_string($package['version'])) {
+                break;
+            }
+
+            return $package['version'];
+        }
+
+        return self::SULU_DEFAULT_VERSION;
     }
 
     /**
      * Read composer.json file and return version of app.
-     *
-     * @param string $dir
-     *
-     * @return string
      */
-    private function getAppVersion($dir)
+    private function getAppVersion(string $dir): ?string
     {
-        $version = null;
-
-        /** @var SplFileInfo $composerFile */
         $composerFile = new SplFileInfo($dir . '/composer.json', '', '');
         if (!$composerFile->isFile()) {
-            return $version;
+            return null;
         }
 
         $composerJson = \json_decode($composerFile->getContents(), true);
-        if (!\array_key_exists('version', $composerJson)) {
-            return $version;
+
+        if (!\is_array($composerJson)
+            || !\array_key_exists('version', $composerJson)
+            || !\is_string($composerJson['version'])
+        ) {
+            return null;
         }
 
         return $composerJson['version'];
