+++ src/Sulu/Bundle/AdminBundle/SuluAdminBundle.php
@@ -14,6 +14,7 @@ namespace Sulu\Bundle\AdminBundle;
 use Sulu\Bundle\AdminBundle\DependencyInjection\Compiler\AddAdminPass;
 use Sulu\Bundle\AdminBundle\DependencyInjection\Compiler\AddMetadataProviderPass;
 use Sulu\Bundle\AdminBundle\DependencyInjection\Compiler\ExposeResourceRoutesPass;
+use Sulu\Bundle\AdminBundle\DependencyInjection\Compiler\SuluVersionPass;
 use Sulu\Component\Symfony\CompilerPass\TaggedServiceCollectorCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
@@ -27,6 +28,7 @@ final class SuluAdminBundle extends Bundle
     {
         parent::build($container);
 
+        $container->addCompilerPass(new SuluVersionPass());
         $container->addCompilerPass(new AddAdminPass());
         $container->addCompilerPass(new AddMetadataProviderPass());
         $container->addCompilerPass(
