+++ src/Sulu/Bundle/AdminBundle/Tests/Unit/DependencyInjection/Compiler/Resources/VersionPass/composer.json
